Rename folder store/extension -> store/identities
