+++ packages/bierzo-wallet/src/store/balances/index.unit.spec.ts
@@ -4,7 +4,7 @@ import { disconnect } from "../../logic/connection";
 import { getExtensionStatus } from "../../logic/extension";
 import { aNewStore } from "../../store";
 import { withChainsDescribe } from "../../utils/test/testExecutor";
-import { setIdentitiesStateAction } from "../extension";
+import { setIdentitiesStateAction } from "../identities";
 import { addBalancesAction, getBalances } from "./actions";
 
 withChainsDescribe("Tokens reducer", () => {
